requirements.txt
ocrmac
google-genai
python-dotenv
requests==2.32.5

